Create Item Model
* Generate Item model
* Create migration for items
* Generate a seeds/development.rb file using seed_dump gem and the sample sqlite database provided.
This allows us to generate the sample data in the database by running rake db:seed. This command uses the rails environment to decide which seed file to run. In this case we only want to seed the database on development and hence we only have a development.rb seeds file in the db/seeds directoy.