Add /search Endpoint
Used controller called SearchItemsController to preserve the REST guidelines of using the index action/method. We could quite easily have added a search action to the ItemsController (or used the index action on that controller). That way would make sense to have url format of either /items/search or simply /items, we could of course map /search to items#search or items#index action but I chose instead to seperate item searching to it's own controller as I feel it conforms better to the REST approach and I don't particularly like mapping custom endpoints to resource controllers.

* Additionally created helper methods for handling JSON in controller responses and in tests