Add ItemSearch Model
ItemSearch model handles building up the query using the item model scopes. It determines which set of scopes to include based on the params. Once it has a record set of 100, it then uses ruby to sort and limit the number of records based on a score. This score is calculated by the RelevanceEvaluator.